[Refactor] Rename PD connector files for clarity
- monkey_patch.py → connector_installer.py (installs the connector override)
- patched_mooncake_connector.py → mooncake_pd_adapter.py (adapts for PD disaggregation)
- Update all import references in omni_stage.py, test_pd_disaggregation.py, and docs
